Working simplistic 2-way sinc with Firestore
It's not very efficient, I think, as it seems to reload
the entire collection each time, but it's probably
fine for now.

The mechanism for getting the components to re-render when
Firestore reports new data is very ropy, though. It
requires having both the backendStore refs and the
components' local state objects being referenced in
the HTML template.